flatten: mirrored_text widget hierarchy
