+++ packages/client/lib/app/core/widgets/mirrored_text/mobx/mirrored_text_store.dart
