--- packages/client/lib/app/core/widgets/mirrored_text/stack/constants/constants.dart
@@ -1,1 +1,0 @@
-export 'types/types.dart';
