--- packages/client/lib/app/core/widgets/mirrored_text/stack/presentation/presentation.dart
@@ -1,2 +1,0 @@
-export 'mobx/mirrored_text_store.dart';
-export 'widget/mirrored_text.dart';
