--- packages/client/lib/app/core/widgets/mirrored_text/stack/stack.dart
@@ -1,2 +1,0 @@
-export 'constants/constants.dart';
-export 'presentation/presentation.dart';
